# Explore new pipeline that overlaps optimizer with emb_lookup